CsvSigil (csv_sigil v0.1.0)

Functions

sigil_v(lines, list)

Implement the ~v sigil, which takes a string containing multiple comma-separated lines, converts each line into a list, and returns a list of those lists. Numeric numbers are converted into float numbers.

With an h option, the sigil returns a list of keyword lists. The keys in each keyword list is taken from the first line. (See the second example below.)

Example usage

  iex> import CsvSigil
  iex> ~v"""
  ...> 1,2,3
  ...> cat,dog
  ...> """
  [[1.0, 2.0, 3.0], ["cat", "dog"]]

  iex> import CsvSigil
  iex> ~v"""
  ...> Item,Qty,Price
  ...> Teddy bear,4,34.95
  ...> Milk,1,2.99
  ...> Battery,6,8.00
  ...> """h
  [
    [Item: "Teddy bear", Qty: 4.0, Price: 34.95],
    [Item: "Milk", Qty: 1.0, Price: 2.99],
    [Item: "Battery", Qty: 6.0, Price: 8.00]
  ]